feat: begun work on paragraph detection
Added a method to listen to window scrolling events, and get the DOM objects which are currently in the viewport. At the moment, it's using a naive approach which iterates over each DOM element whenever the user scrolls, so it's probably really inefficient, but it's fine for now. 

I also updated the formatting in the references file.