feat: Add support for logging scroll events in Firefox extension
This commit adds the ability to log scroll events in the Firefox extension. It introduces a new function `logScrollEvents` that calls `getVisibleTextElements` to get the elements currently in the viewport and logs their count. The function is then attached as an event listener to the `scroll` event.

The code changes include:
- Added `logScrollEvents` function
- Attached `logScrollEvents` as an event listener for the `scroll` event

BREAKING CHANGE: Removed unused file getTextElements.js