feat: Remove Chrome extension support and add cleanup scripts
This commit removes the Chrome extension files and manifest, as well as renames and moves some Firefox backup files. Additionally, it adds two new scripts for cleaning up and building the project.

BREAKING CHANGE: The Chrome extension support has been removed due to difficulties with manifest v3. However, as plasmo is browser-agnostic, this shouldn't pose much of an issue